userdash completed almost

+++ src/controllers/orderController.ts
@@ -4,12 +4,13 @@ import Order from "../models/orderModel";
 import Cart from "../models/cartModel";
 import { CustomError } from "../middlewares/error";
 import checkPermission from "../helpers/authHelper";
-import { Types } from "mongoose";
 
 // Initialize Stripe
 const stripe = new Stripe(process.env.STRIPE_SECRET_KEY!, {
-  apiVersion: "2024-11-20.acacia"
+  apiVersion: "2024-11-20.acacia",
 });
+
+// *********************** ADMIN ***************************
 
 // *********************************************************
 // ************ Create an Order and Initiate Payment *******
@@ -35,25 +36,34 @@ export const createOrder = async (
     }
     const discountedPrice = totalPrice - couponDiscount;
 
+    // Generate order number based on current date and time
+    const now = new Date();
+    const orderNumber = `${String(now.getDate()).padStart(2, "0")}${String(
+      now.getMonth() + 1
+    ).padStart(2, "0")}${String(now.getFullYear()).slice(-2)}${String(
+      now.getHours()
+    ).padStart(2, "0")}${String(now.getMinutes()).padStart(2, "0")}`;
+
     const session = await stripe.checkout.sessions.create({
       payment_method_types: ["card"],
-      line_items: cart.products.map(product => ({
+      line_items: cart.products.map((product) => ({
         price_data: {
           currency: "usd",
           product_data: {
-            name: product.productName
+            name: product.productName,
           },
-          unit_amount: Math.round(product.totalPrice * 100)
+          unit_amount: Math.round(product.pricePerUnit * 100),
         },
-        quantity: product.quantity
+        quantity: product.quantity,
       })),
       mode: "payment",
       success_url: `${process.env.LOCAL_DOMAIN}/success`,
       cancel_url: `${process.env.LOCAL_DOMAIN}/cancel`,
-      metadata: { userId, couponId }
+      metadata: { userId, couponId },
     });
 
     const newOrder = new Order({
+      orderNumber: orderNumber,
       userId,
       products: cart.products,
       totalQuantity: cart.totalQuantity,
@@ -66,7 +76,7 @@ export const createOrder = async (
       paymentUrl: session.url,
       status: "Pending",
       paymentStatus: "Unpaid",
-      paymentMethod
+      paymentMethod,
     });
 
     await newOrder.save();
@@ -74,7 +84,7 @@ export const createOrder = async (
     res.status(201).json({
       message: "Order created successfully.",
       data: newOrder,
-      sessionId: session.id
+      sessionId: session.id,
     });
   } catch (error) {
     if (error instanceof Error) {
@@ -120,7 +130,7 @@ export const stripeWebhook = async (req: Request, res: Response) => {
           {
             status: "Confirmed",
             paymentStatus: "Paid",
-            paymentDate: Date.now()
+            paymentDate: Date.now(),
           },
           { new: true }
         );
@@ -137,35 +147,35 @@ export const stripeWebhook = async (req: Request, res: Response) => {
 
       // Create a new checkout session if expired
       const cart = await Cart.findOne({
-        userId: expiredSession.metadata.userId
+        userId: expiredSession.metadata.userId,
       });
       if (cart && cart.products.length > 0) {
         const newSession = await stripe.checkout.sessions.create({
           payment_method_types: ["card"],
-          line_items: cart.products.map(product => ({
+          line_items: cart.products.map((product) => ({
             price_data: {
               currency: "usd",
               product_data: {
-                name: product.productName
+                name: product.productName,
               },
-              unit_amount: Math.round(product.totalPrice * 100)
+              unit_amount: Math.round(product.totalPrice * 100),
             },
-            quantity: product.quantity
+            quantity: product.quantity,
           })),
           mode: "payment",
           success_url: `${process.env.LOCAL_DOMAIN}/success`,
           cancel_url: `${process.env.LOCAL_DOMAIN}/cancel`,
           metadata: {
             userId: expiredSession.metadata.userId,
-            couponId: expiredSession.metadata.couponId
-          }
+            couponId: expiredSession.metadata.couponId,
+          },
         });
 
         await Order.findOneAndUpdate(
           { paymentId: expiredSession.id },
           {
             paymentUrl: newSession.url,
-            paymentId: newSession.id
+            paymentId: newSession.id,
           },
           { new: true }
         );
@@ -179,7 +189,7 @@ export const stripeWebhook = async (req: Request, res: Response) => {
 };
 
 // *********************************************************
-// ************ Retrieve All Orders for a User ************
+// ************ Retrieve All Orders for Admin **************
 // *********************************************************
 export const getOrders = async (
   req: Request,
@@ -188,7 +198,24 @@ export const getOrders = async (
 ) => {
   try {
     const { userId } = req;
-    const orders = await Order.find({ userId }).sort({ createdAt: -1 });
+    const { orderId } = req.query;
+
+    // Check Permission
+    const permissionCheck = await checkPermission(userId, "orders", 2);
+    if (!permissionCheck) {
+      return res.status(403).json({ message: "Permission denied" });
+    }
+
+    let orders;
+    if (orderId) {
+      orders = await Order.findById(orderId);
+      if (!orders) {
+        return res.status(404).json({ message: "Order not found." });
+      }
+    } else {
+      orders = await Order.find().sort({ createdAt: -1 }).populate("userId");
+    }
+
     res.status(200).json({ data: orders });
   } catch (error) {
     next(error);
@@ -196,42 +223,27 @@ export const getOrders = async (
 };
 
 // *********************************************************
-// ************ Retrieve a Single Order by ID **************
+// ************ Update Order (Admin) ***********************
 // *********************************************************
-export const getOrderById = async (
+export const updateOrder = async (
   req: Request,
   res: Response,
   next: NextFunction
 ) => {
   try {
-    const { orderId } = req.params;
-    const order = await Order.findById(orderId);
-    if (!order) {
-      return res.status(404).json({ message: "Order not found." });
+    const { userId } = req;
+    const { orderId } = req.query;
+    const updateData = req.body;
+
+    // Check Permission
+    const permissionCheck = await checkPermission(userId, "orders", 3);
+    if (!permissionCheck) {
+      return res.status(403).json({ message: "Permission denied" });
     }
-    res.status(200).json({ data: order });
-  } catch (error) {
-    next(error);
-  }
-};
 
-// *********************************************************
-// ************ Update Order Status (e.g., Cancel) *********
-// *********************************************************
-export const updateOrderStatus = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
-  try {
-    const { orderId } = req.params;
-    const { status, paymentStatus } = req.body;
-
-    const updatedOrder = await Order.findByIdAndUpdate(
-      orderId,
-      { status, paymentStatus },
-      { new: true }
-    );
+    const updatedOrder = await Order.findByIdAndUpdate(orderId, updateData, {
+      new: true,
+    });
 
     if (!updatedOrder) {
       return res.status(404).json({ message: "Order not found." });
@@ -254,7 +266,14 @@ export const deleteOrder = async (
   next: NextFunction
 ) => {
   try {
-    const { orderId } = req.params;
+    const { userId } = req;
+    const { orderId } = req.query;
+
+    // Check Permission
+    const permissionCheck = await checkPermission(userId, "orders", 4);
+    if (!permissionCheck) {
+      return res.status(403).json({ message: "Permission denied" });
+    }
 
     const deletedOrder = await Order.findByIdAndDelete(orderId);
     if (!deletedOrder) {
@@ -262,6 +281,36 @@ export const deleteOrder = async (
     }
 
     res.status(200).json({ message: "Order deleted successfully." });
+  } catch (error) {
+    next(error);
+  }
+};
+
+// *********************** USER ****************************
+
+// *********************************************************
+// ************ Retrieve Orders by User ID *****************
+// *********************************************************
+export const getOrdersByUserId = async (
+  req: Request,
+  res: Response,
+  next: NextFunction
+) => {
+  try {
+    const { userId } = req;
+    const { orderId } = req.query;
+
+    let orders;
+    if (orderId) {
+      orders = await Order.findOne({ userId, _id: orderId });
+      if (!orders) {
+        return res.status(404).json({ message: "Order not found." });
+      }
+    } else {
+      orders = await Order.find({ userId }).sort({ createdAt: -1 });
+    }
+
+    res.status(200).json({ data: orders });
   } catch (error) {
     next(error);
   }
